added weapon data.
