+++ FF8/FF8.csproj
@@ -157,6 +157,7 @@
     <Compile Include="Menu\Kernel_bin.Gender.cs" />
     <Compile Include="Menu\Kernel_bin.Junctionable_GFs_Data.cs" />
     <Compile Include="Menu\Kernel_bin.Magic_Data.cs" />
+    <Compile Include="Menu\Kernel_bin.Weapons_Data.cs" />
     <Compile Include="Menu\Magazine.cs" />
     <Compile Include="Menu\Magazine.InsertCustomEntries.cs" />
     <Compile Include="Menu\Magazine.ID.cs" />
