+++ FF8/Menu/Kernel_bin.cs
@@ -80,37 +80,5 @@ namespace FF8
                 }
             }
         }
-
-        public class Weapons_Data
-        {
-            public const int count =33;
-            public const int id=4;
-            public FF8String Name { get; private set; }
-            public override string ToString() => Name;
-
-            //0x0000	2 bytes Offset to weapon name
-            //0x0002	1 byte Renzokuken finishers
-
-            //0x01 = Rough Divide
-            //0x02 = Fated Circle
-            //0x04 = Blasting Zone
-            //0x08 = Lion Heart
-            //0x0003	1 byte Unknown
-            public Saves.Characters Character;//0x0004	1 byte Character ID
-
-            //0x0005	1 bytes Attack Type
-            //0x0006	1 byte Attack Power
-            //0x0007	1 byte Attack Parameter
-            //0x0008	1 byte STR Bonus
-            //0x0009	1 byte Weapon Tier
-            //0x000A	1 byte Crit Bonus
-            //0x000B	1 byte Melee Weapon?
-
-            internal void Read(BinaryReader br,int string_id = 0)
-            {
-                Name = Memory.Strings.Read(Strings.FileID.KERNEL, id, string_id);
-                br.BaseStream.Seek(2, SeekOrigin.Current);
-            }
-        }
     }
 }
